# Conflicts:
#	routes/web.php

+++ app/Filament/Resources/RoomBookingResource/Pages/EditRoomBooking.php
@@ -49,12 +49,21 @@ class EditRoomBooking extends EditRecord
     {
         $record = $this->record; // Record vừa được tạo
         $data = $this->data; // Data từ form
-        $roomBookingService = new RoomBookingService();
-        // Xóa chi tiết cũ nếu có
-        $roomBookingService->deleteBookingDetails($record->id ?? null);
-        $roomBookingService->createBookingDetails($record, $data, false);
-        // Cập nhật trạng thái xung đột
-        $roomBookingService->updateDuplicateStatus($record->room_id);
+
+        if ($record->status == 'pending') {   
+            $roomBookingService = new RoomBookingService();
+            // Xóa chi tiết cũ nếu có
+            $roomBookingService->deleteBookingDetails($record->id ?? null);
+            $roomBookingService->createBookingDetails($record, $data, false);
+            // Cập nhật trạng thái xung đột
+            $roomBookingService->updateDuplicateStatus($record->room_id);
+        }
+        // Thông báo thành công
+        Notification::make()
+            ->title('Cập nhật thành công')
+            ->body('Đặt phòng đã được cập nhật thành công.')
+            ->success()
+            ->send();
     }
 
     // xóa thông báo mặc định sau khi lưu
